:recycle: [Refactor]: study 도메인 패키지 추가
