+++ src/main/java/timify/com/member/dto/MemberRequest.java
@@ -2,7 +2,6 @@ package timify.com.member.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.Getter;
 
 import java.time.LocalDate;
@@ -32,10 +31,5 @@ public class MemberRequest {
         String loginType;
     }
 
-    @Getter
-    public static class studyTypeInsertRequest {
-        @NotBlank
-        @Size(max = 30)
-        String title;
-    }
+
 }
