+++ src/main/java/timify/com/member/dto/MemberResponse.java
@@ -25,13 +25,4 @@ public class MemberResponse {
         LocalDate birth;
     }
 
-    @Builder
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class studyTypeInsertDto {
-        Long studyTypeId;
-        String studyTypeTitle;
-    }
-
 }
