+++ src/main/java/timify/com/member/repository/StudyTypeRepository.java
@@ -1,7 +1,7 @@
 package timify.com.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import timify.com.domain.StudyType;
+import timify.com.study.domain.StudyType;
 
 public interface StudyTypeRepository extends JpaRepository<StudyType, Long> {
 }
