+++ src/main/java/timify/com/study/StudyConverter.java
@@ -1,0 +1,22 @@
+package timify.com.study;
+
+import timify.com.study.domain.CategoryStatus;
+import timify.com.study.domain.StudyType;
+import timify.com.study.dto.StudyResponse;
+
+public class StudyConverter {
+    public static StudyType toStudyType(String title, int order) {
+        return StudyType.builder()
+                .title(title)
+                .order_num(order)
+                .status(CategoryStatus.ACTIVE)
+                .build();
+    }
+
+    public static StudyResponse.studyTypeInsertDto toStudyTypeInsertDto(StudyType studyType) {
+        return StudyResponse.studyTypeInsertDto.builder()
+                .studyTypeId(studyType.getId())
+                .studyTypeTitle(studyType.getTitle())
+                .build();
+    }
+}
