+++ src/main/java/timify/com/study/dto/StudyRequest.java
@@ -1,0 +1,14 @@
+package timify.com.study.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.Getter;
+
+public class StudyRequest {
+    @Getter
+    public static class studyTypeInsertRequest {
+        @NotBlank
+        @Size(min = 1, max = 30)
+        String title;
+    }
+}
