+++ src/main/java/timify/com/study/dto/StudyResponse.java
@@ -1,0 +1,17 @@
+package timify.com.study.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+public class StudyResponse {
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class studyTypeInsertDto {
+        Long studyTypeId;
+        String studyTypeTitle;
+    }
+}
